--- src/backend/ci/core/common/common-es/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.tencent.devops.common.es.ESAutoConfiguration
