+++ src/backend/ci/core/log/biz-log/src/main/kotlin/com/tencent/devops/log/configuration/LogCommonConfiguration.kt
@@ -24,12 +24,39 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.common.es
+package com.tencent.devops.log.configuration
 
-import org.elasticsearch.client.RestHighLevelClient
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.boot.autoconfigure.AutoConfigureOrder
+import org.springframework.boot.context.properties.EnableConfigurationProperties
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.core.Ordered
 
-data class ESClient(
-    val name: String,
-    val client: RestHighLevelClient,
-    val mainCluster: Boolean? = false
-)
+@Configuration
+@AutoConfigureOrder(Ordered.HIGHEST_PRECEDENCE)
+class LogCommonConfiguration {
+
+    @Value("\${log.storage.type:#{null}}")
+    private val type: String? = null
+
+    @Bean
+    fun storageProperties(): StorageProperties {
+        if (type.isNullOrBlank()) {
+            throw IllegalArgumentException("storage type of build log didn't config: log.storage.type, it must be either of 'lucene' or 'elasticsearch'.")
+        }
+        return StorageProperties()
+    }
+
+    @Bean
+    fun defaultKeywords() = listOf(
+        "error ( )",
+        "Scripts have compiler errors",
+        "fatal error",
+        "no such",
+        // "Exception :",;
+        "Code Sign error",
+        "BUILD FAILED",
+        "Failed PVR compression"
+    )
+}
