+++ src/backend/ci/core/log/biz-log/src/main/kotlin/com/tencent/devops/log/cron/impl/IndexCleanJobESImpl.kt
@@ -24,13 +24,14 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.log.cron
+package com.tencent.devops.log.cron.impl
 
 import com.tencent.devops.common.api.exception.OperationException
 import com.tencent.devops.common.redis.RedisLock
 import com.tencent.devops.common.redis.RedisOperation
 import com.tencent.devops.log.client.LogClient
-import com.tencent.devops.log.util.IndexNameUtils.LOG_PREFIX
+import com.tencent.devops.log.configuration.StorageProperties
+import com.tencent.devops.log.cron.IndexCleanJob
 import org.elasticsearch.action.admin.indices.close.CloseIndexRequest
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest
 import org.elasticsearch.client.indices.GetIndexRequest
@@ -38,26 +39,28 @@ import org.elasticsearch.client.RequestOptions
 import org.elasticsearch.client.RestHighLevelClient
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
 import java.time.temporal.ChronoUnit
 
 @Component
-class ESIndexCloseJob @Autowired constructor(
+@ConditionalOnProperty(prefix = "log.storage", name = ["type"], havingValue = "elasticsearch")
+class IndexCleanJobESImpl @Autowired constructor(
+    private val storageProperties: StorageProperties,
     private val client: LogClient,
     private val redisOperation: RedisOperation
-) {
+) : IndexCleanJob {
 
-    private var closeIndexInDay = 30 // default is expire in 30 days
-    private var deleteIndexInDay = 90 // default be deleted in 90 days
+    private var closeIndexInDay = storageProperties.closeInDay
+    private var deleteIndexInDay = storageProperties.deleteInDay
 
     /**
      * 2 am every day
      */
     @Scheduled(cron = "0 0 2 * * ?")
-    fun closeIndex() {
+    override fun closeIndex() {
         logger.info("Start to close index")
         val redisLock = RedisLock(redisOperation, ES_INDEX_CLOSE_JOB_KEY, 20)
         try {
@@ -74,7 +77,7 @@ class ESIndexCloseJob @Autowired constructor(
         }
     }
 
-    fun updateExpireIndexDay(expired: Int) {
+    override fun updateExpireIndexDay(expired: Int) {
         logger.warn("Update the expire index day from $expired to ${this.closeIndexInDay}")
         if (expired <= 10) {
             logger.warn("The expired is illegal")
@@ -83,7 +86,7 @@ class ESIndexCloseJob @Autowired constructor(
         this.closeIndexInDay = expired
     }
 
-    fun getExpireIndexDay() = closeIndexInDay
+    override fun getExpireIndexDay() = closeIndexInDay
 
     private fun closeESIndexes() {
         client.getActiveClients().forEach { c ->
@@ -115,18 +118,18 @@ class ESIndexCloseJob @Autowired constructor(
 
     private fun deleteESIndexes() {
         client.getActiveClients().forEach { c ->
-            val indexes = c.client
+            val response = c.client
                 .indices()
                 .get(GetIndexRequest(), RequestOptions.DEFAULT)
 
-            if (indexes.indices.isEmpty()) {
+            if (response.indices.isEmpty()) {
                 return
             }
 
             val deathLine = LocalDateTime.now()
                 .minus(deleteIndexInDay.toLong(), ChronoUnit.DAYS)
             logger.info("Get the death line - ($deathLine)")
-            indexes.indices.forEach { index ->
+            response.indices.forEach { index ->
                 if (expire(deathLine, index)) {
                     deleteESIndex(c.client, index)
                 }
@@ -141,27 +144,8 @@ class ESIndexCloseJob @Autowired constructor(
         logger.info("Get the delete es response - ${resp.isAcknowledged}")
     }
 
-    private fun expire(deathLine: LocalDateTime, index: String): Boolean {
-        try {
-            if (!index.startsWith(LOG_PREFIX)) {
-                return false
-            }
-            val dateStr = index.replace(LOG_PREFIX, "") + " 00:00"
-            val format = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm")
-            val date = LocalDateTime.parse(dateStr, format)
-
-            if (deathLine > date) {
-                logger.info("[$index] The index is expire ($deathLine|$date)")
-                return true
-            }
-        } catch (t: Throwable) {
-            logger.warn("[$index] Fail to check if the index expire", t)
-        }
-        return false
-    }
-
     companion object {
-        private val logger = LoggerFactory.getLogger(ESIndexCloseJob::class.java)
+        private val logger = LoggerFactory.getLogger(IndexCleanJobESImpl::class.java)
         private const val ES_INDEX_CLOSE_JOB_KEY = "log:es:index:close:job:lock:key"
     }
 }
