+++ src/backend/ci/core/log/biz-log/src/main/kotlin/com/tencent/devops/log/es/ESProperties.kt
@@ -24,28 +24,13 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.log.configuration
+package com.tencent.devops.log.es
 
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.scheduling.annotation.EnableAsync
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor
+import org.springframework.boot.context.properties.ConfigurationProperties
 
-/**
- *
- * Powered By Tencent
- */
-@Configuration
-@EnableAsync
-class ThreadConfig {
-
-    @Bean
-    fun threadPoolTaskExecutor() = ThreadPoolTaskExecutor()
-
-    @Bean
-    fun logMessagesThreadPoolTaskExecutor() = ThreadPoolTaskExecutor().apply {
-        corePoolSize = 10
-        maxPoolSize = 80
-        setQueueCapacity(1000)
-    }
-}
+@ConfigurationProperties(prefix = "log.elasticsearch")
+data class ESProperties(
+    val ip: String? = null,
+    val port: Int? = 0,
+    val cluster: String? = null
+)
