+++ src/backend/ci/core/log/biz-log/src/main/kotlin/com/tencent/devops/log/lucene/LuceneProperties.kt
@@ -24,23 +24,12 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.log.configuration
+package com.tencent.devops.log.lucene
 
-import com.tencent.devops.common.es.ESClient
-import com.tencent.devops.log.client.impl.LogClientImpl
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.autoconfigure.AutoConfigureOrder
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.core.Ordered
+import org.springframework.boot.context.properties.ConfigurationProperties
 
-@Configuration
-@ConditionalOnWebApplication
-@AutoConfigureOrder(Ordered.LOWEST_PRECEDENCE)
-class LogClientConfiguration {
-
-    @Bean
-    fun logClient(@Autowired transportClient: ESClient) =
-        LogClientImpl(transportClient)
-}
+@ConfigurationProperties(prefix = "log.lucene")
+data class LuceneProperties(
+    val dataDirectory: String? = null,
+    val indexMaxSize: Int? = Int.MAX_VALUE
+)
